+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Changes
 
-* 2018-03-02 1.2.2 Update for Rails 5
+* 2018-03-02 1.3.0 Add columns e.g. for geolocation and dba
+* 2018-03-01 1.2.0 Update for Ruby 2 & Rails 5
 * 2015-07-07 1.0.9 Update gemspec to use file manifest
 * 2012-03-14 1.0.8 Update docs, tests